requirements.txt
click
packaging
pytenable
